:sparkles: Implemented timer interface for callback functions
- Added timer interface for stm32f1 series
- Added a working demo for the stm32f1 timer
- Added generic timer helper/interface for STM32 MCUs
- Updated general_purpose_timer to use a manager
Note: This is an unpolished, but working commit. Will improve it.

+++ src/stm32_generic/timer.cpp
@@ -62,11 +62,11 @@ void timer::schedule(hal::time_duration p_delay, u32 p_timer_clock_frequency)
 
   u32 const time_per_tick_ns = scale_factor_ns / p_timer_clock_frequency;
   u32 const ticks_required = p_delay.count() / time_per_tick_ns;
-  u32 const prescaler_value = ticks_required / timer_max_ticks;
+  u16 const prescaler_value = ticks_required / timer_max_ticks;
   u32 const prescaler_frequency =
     p_timer_clock_frequency / (prescaler_value + 1);
   u32 const prescaler_time_per_tick_ns = scale_factor_ns / prescaler_frequency;
-  u32 const prescaler_ticks_required =
+  u16 const prescaler_ticks_required =
     p_delay.count() / prescaler_time_per_tick_ns;
 
   bit_modify(reg->prescale_register).insert<prescaler>(prescaler_value);
